clementine_extended_rpc

Function get_fee_rate_from_mempool_space

pub async fn get_fee_rate_from_mempool_space(
    url: &Option<String>,
    endpoint: &Option<String>,
    network: Network,
) -> Result<Amount, BitcoinRPCError>

Fetches the current recommended fee rate from the provider. Currently only supports Mempool Space API. This function is used to get the fee rate in sat/vkb (satoshis per kilovbyte). See Mempool Space API for more details.